Move generic backendData utils to separate file
