+++ ui/src/components/management/PermitAccordionHeading.tsx
@@ -5,7 +5,8 @@ import { useDispatch } from "react-redux";
 import { IonButton, IonCol, IonGrid, IonRow, IonText } from "@ionic/react";
 import Moment from "react-moment";
 import IPermit from "../../interfaces/IPermit";
-import { getRouteTransportsOfPermit, onRetry } from "../../utils/managementBackendData";
+import { onRetry } from "../../utils/backendData";
+import { getRouteTransportsOfPermit } from "../../utils/managementBackendData";
 import { DATE_FORMAT } from "../../utils/constants";
 
 interface PermitAccordionHeadingProps {
