+++ ui/src/components/management/RouteTransportInfo.tsx
@@ -9,7 +9,8 @@ import IPermit from "../../interfaces/IPermit";
 import IRoute from "../../interfaces/IRoute";
 import IRouteTransport from "../../interfaces/IRouteTransport";
 import ISupervisor from "../../interfaces/ISupervisor";
-import { createRouteTransport, deleteRouteTransport, onRetry, updateRouteTransport } from "../../utils/managementBackendData";
+import { onRetry } from "../../utils/backendData";
+import { createRouteTransport, deleteRouteTransport, updateRouteTransport } from "../../utils/managementBackendData";
 import { DATE_FORMAT, TransportStatus } from "../../utils/constants";
 import BridgeGrid from "./BridgeGrid";
 import RouteInfoGrid from "./RouteInfoGrid";
