+++ ui/src/pages/BridgeDetail.tsx
@@ -13,7 +13,8 @@ import TrafficSupervisorsAccordion from "../components/TrafficSupervisorsAccordi
 import IPermit from "../interfaces/IPermit";
 import IRouteBridge from "../interfaces/IRouteBridge";
 import ISupervision from "../interfaces/ISupervision";
-import { getSupervision, onRetry, updateConformsToPermit } from "../utils/supervisionBackendData";
+import { onRetry } from "../utils/backendData";
+import { getSupervision, updateConformsToPermit } from "../utils/supervisionBackendData";
 
 interface BridgeDetailProps {
   supervisionId: string;
