+++ ui/src/pages/Home.tsx
@@ -6,7 +6,8 @@ import type { SegmentChangeEventDetail } from "@ionic/core";
 import { IonContent, IonLabel, IonPage, IonSegment, IonSegmentButton, IonSlide, IonSlides } from "@ionic/react";
 import Header from "../components/Header";
 import { useTypedSelector } from "../store/store";
-import { getCompanyTransportsList, getSupervisionList, onRetry } from "../utils/supervisionBackendData";
+import { onRetry } from "../utils/backendData";
+import { getCompanyTransportsList, getSupervisionList } from "../utils/supervisionBackendData";
 import SupervisionList from "../components/SupervisionList";
 import "./Home.css";
 import CompanyTransportsAccordion from "../components/CompanyTransportsAccordion";
