+++ ui/src/pages/Photos.tsx
@@ -8,7 +8,8 @@ import { useHistory, useParams } from "react-router-dom";
 import moment from "moment";
 import { v4 as uuidv4 } from "uuid";
 import camera from "../theme/icons/camera_white.svg";
-import { deleteImage, getSupervision, onRetry, sendImageUpload } from "../utils/supervisionBackendData";
+import { onRetry } from "../utils/backendData";
+import { deleteImage, getSupervision, sendImageUpload } from "../utils/supervisionBackendData";
 import { DATE_TIME_FORMAT } from "../utils/constants";
 import { getOrigin } from "../utils/request";
 import ImagePreview from "../components/ImagePreview";
