+++ ui/src/pages/RouteTransportDetail.tsx
@@ -11,7 +11,8 @@ import PermitTransportAccordion from "../components/PermitTransportAccordion";
 import IPermit from "../interfaces/IPermit";
 import IRoute from "../interfaces/IRoute";
 import { useTypedSelector } from "../store/store";
-import { getRouteTransportOfSupervisor, onRetry } from "../utils/supervisionBackendData";
+import { onRetry } from "../utils/backendData";
+import { getRouteTransportOfSupervisor } from "../utils/supervisionBackendData";
 import BridgeCardList from "../components/BridgeCardList";
 import { filterFinishedSupervisions } from "../utils/supervisionUtil";
 import ISupervision from "../interfaces/ISupervision";
