+++ ui/src/pages/Supervision.tsx
@@ -12,7 +12,8 @@ import SupervisionObservations from "../components/SupervisionObservations";
 import SupervisionPhotos from "../components/SupervisionPhotos";
 import ISupervision from "../interfaces/ISupervision";
 import { useTypedSelector } from "../store/store";
-import { cancelSupervision, deleteSupervisionImages, getSupervision, onRetry, updateSupervisionReport } from "../utils/supervisionBackendData";
+import { onRetry } from "../utils/backendData";
+import { cancelSupervision, deleteSupervisionImages, getSupervision, updateSupervisionReport } from "../utils/supervisionBackendData";
 import ISupervisionReport from "../interfaces/ISupervisionReport";
 import { SupervisionStatus } from "../utils/constants";
 import { reportHasUnsavedChanges } from "../utils/supervisionUtil";
