+++ ui/src/pages/SupervisionSummary.tsx
@@ -11,7 +11,8 @@ import SupervisionObservationsSummary from "../components/SupervisionObservation
 import SupervisionPhotos from "../components/SupervisionPhotos";
 import ISupervision from "../interfaces/ISupervision";
 import { useTypedSelector } from "../store/store";
-import { finishSupervision, getSupervision, onRetry } from "../utils/supervisionBackendData";
+import { onRetry } from "../utils/backendData";
+import { finishSupervision, getSupervision } from "../utils/supervisionBackendData";
 import SupervisionFooter from "../components/SupervisionFooter";
 import { SupervisionStatus } from "../utils/constants";
