+++ ui/src/utils/supervisionBackendData.ts
@@ -1,8 +1,6 @@
 import type { Dispatch } from "redux";
 import ISupervisionImage from "../interfaces/ISupervisionImage";
 import ISupervisionImageInput from "../interfaces/ISupervisionImageInput";
-import IRoute from "../interfaces/IRoute";
-import IRouteBridge from "../interfaces/IRouteBridge";
 import ISupervision from "../interfaces/ISupervision";
 import ISupervisionReport from "../interfaces/ISupervisionReport";
 import { getOrigin } from "./request";
@@ -10,35 +8,8 @@ import { actions as supervisionActions } from "../store/supervisionSlice";
 import ICompanyTransports from "../interfaces/ICompanyTransports";
 import IRouteTransport from "../interfaces/IRouteTransport";
 import IDenyCrossingInput from "../interfaces/IDenyCrossingInput";
-import IUserData from "../interfaces/IUserData";
 
 const notOkError = "Network response was not ok";
-
-export const onRetry = (failureCount: number, error: string): boolean => {
-  // Retry forever by returning true
-  console.error("ERROR", failureCount, error);
-  return true;
-};
-
-export const getSupervisorUser = async (dispatch: Dispatch): Promise<IUserData> => {
-  try {
-    console.log("getSupervisorUser");
-    dispatch({ type: supervisionActions.SET_FAILED_QUERY, payload: { getSupervisorUser: false } });
-
-    const userDataResponse = await fetch(`${getOrigin()}/api/ui/userdata`);
-
-    if (userDataResponse.ok) {
-      const userData = (await userDataResponse.json()) as Promise<IUserData>;
-      return await userData;
-    } else {
-      dispatch({ type: supervisionActions.SET_FAILED_QUERY, payload: { getSupervisorUser: true } });
-      throw new Error(notOkError);
-    }
-  } catch (err) {
-    dispatch({ type: supervisionActions.SET_FAILED_QUERY, payload: { getSupervisorUser: true } });
-    throw new Error(err as string);
-  }
-};
 
 export const getCompanyTransportsList = async (dispatch: Dispatch): Promise<ICompanyTransports[]> => {
   try {
@@ -78,46 +49,6 @@ export const getRouteTransportOfSupervisor = async (routeTransportId: number, di
     }
   } catch (err) {
     dispatch({ type: supervisionActions.SET_FAILED_QUERY, payload: { getRouteTransport: true } });
-    throw new Error(err as string);
-  }
-};
-
-export const getRouteGeometry = async (routeId: number, dispatch: Dispatch): Promise<IRoute> => {
-  try {
-    console.log("GetRouteGeometry", routeId);
-    dispatch({ type: supervisionActions.SET_FAILED_QUERY, payload: { getRoute: false } });
-
-    const routeResponse = await fetch(`${getOrigin()}/api/route/getroute?routeId=${routeId}`);
-
-    if (routeResponse.ok) {
-      const route = (await routeResponse.json()) as Promise<IRoute>;
-      return await route;
-    } else {
-      dispatch({ type: supervisionActions.SET_FAILED_QUERY, payload: { getRoute: true } });
-      throw new Error(notOkError);
-    }
-  } catch (err) {
-    dispatch({ type: supervisionActions.SET_FAILED_QUERY, payload: { getRoute: true } });
-    throw new Error(err as string);
-  }
-};
-
-export const getRouteBridgeGeometry = async (routeBridgeId: number, dispatch: Dispatch): Promise<IRouteBridge> => {
-  try {
-    console.log("getRouteBridgeGeometry", routeBridgeId);
-    dispatch({ type: supervisionActions.SET_FAILED_QUERY, payload: { getRouteBridge: false } });
-
-    const routeBridgeResponse = await fetch(`${getOrigin()}/api/routebridge/getroutebridge?routeBridgeId=${routeBridgeId}`);
-
-    if (routeBridgeResponse.ok) {
-      const routeBridge = (await routeBridgeResponse.json()) as Promise<IRouteBridge>;
-      return await routeBridge;
-    } else {
-      dispatch({ type: supervisionActions.SET_FAILED_QUERY, payload: { getRouteBridge: true } });
-      throw new Error(notOkError);
-    }
-  } catch (err) {
-    dispatch({ type: supervisionActions.SET_FAILED_QUERY, payload: { getRouteBridge: true } });
     throw new Error(err as string);
   }
 };
